pyproject.toml
[tool.poetry]
name = "anime-api"
version = "0.15.0"
description = "A collection of wrappers for anime-related APIs"
authors = ["Neki <84998222+Nekidev@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "anime_api"}]

[tool.poetry.dependencies]
python = ">=3.7,<4"
requests = "^2.28.1"
pykakasi = "^2.2.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
